Add 64 bit floats, make more idiomatic.

+++ 2d/64bit/vec/vector.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Vector2D struct {
-	X float32
-	Y float32
+	X float64
+	Y float64
 }
 
-func FromScalar(v float32) Vector2D {
+func FromScalar(v float64) Vector2D {
 	return Vector2D{v, v}
 }
 
@@ -41,11 +41,11 @@ func (v Vector2D) Divide(v2 Vector2D) Vector2D {
 	return Vector2D{v.X / v2.X, v.Y / v2.Y}
 }
 
-func (v Vector2D) MultiplyScalar(s float32) Vector2D {
+func (v Vector2D) MultiplyScalar(s float64) Vector2D {
 	return Vector2D{v.X * s, v.Y * s}
 }
 
-func (v Vector2D) DivideScalar(s float32) Vector2D {
+func (v Vector2D) DivideScalar(s float64) Vector2D {
 	return Vector2D{v.X / s, v.Y / s}
 }
